Small code cleanup.

+++ include/ClusterInterface.hh
@@ -117,7 +117,7 @@ public:
   virtual kinetic::KineticStatus range(
     const std::shared_ptr<const std::string>& start_key,
     const std::shared_ptr<const std::string>& end_key,
-    int maxRequested,
+    size_t maxRequested,
     std::unique_ptr< std::vector<std::string> >& keys) = 0;
   
   //----------------------------------------------------------------------------
