+++ include/KineticCluster.hh
@@ -59,7 +59,7 @@ public:
   kinetic::KineticStatus range(
       const std::shared_ptr<const std::string>& start_key,
       const std::shared_ptr<const std::string>& end_key,
-      int maxRequested,
+      size_t maxRequested,
       std::unique_ptr< std::vector<std::string> >& keys
   );
