+++ test/DataCacheTest.cc
@@ -72,7 +72,7 @@ public:
   kinetic::KineticStatus range(
     const std::shared_ptr<const std::string>& start_key,
     const std::shared_ptr<const std::string>& end_key,
-    int maxRequested,
+    size_t maxRequested,
     std::unique_ptr< std::vector<std::string> >& keys)
   {
     return KineticStatus(StatusCode::OK, "");
